+++ plone/app/event/portlets/portlet_events.py
@@ -1,9 +1,11 @@
 from datetime import datetime
 from Acquisition import aq_inner
+from Products.CMFCore.utils import getToolByName
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from plone.app.layout.navigation.root import getNavigationRootObject
 from plone.app.portlets.cache import render_cachekey
 from plone.app.portlets.portlets import base
+from plone.app.vocabularies.catalog import SearchableTextSourceBinder
 from plone.memoize import ram
 from plone.memoize.compress import xhtml_compress
 from plone.memoize.instance import memoize
@@ -13,10 +15,10 @@ from zope.component import getMultiAdapter
 from zope.formlib import form
 from zope.interface import implements
 
+from plone.event.interfaces import IEventAccessor
 from plone.app.event.base import get_occurrences
 from plone.app.event.base import get_portal_events
 from plone.app.event.base import localized_now
-from plone.event.interfaces import IEventAccessor
 
 from plone.app.portlets import PloneMessageFactory as _
 
@@ -24,28 +26,43 @@ from plone.app.portlets import PloneMessageFactory as _
 class IEventsPortlet(IPortletDataProvider):
 
     count = schema.Int(title=_(u'Number of items to display'),
-                       description=_(u'How many items to list.'),
-                       required=True,
-                       default=5)
+        description=_(u'How many items to list.'),
+        required=True,
+        default=5)
 
     state = schema.Tuple(title=_(u"Workflow state"),
-                         description=_(u"Items in which workflow state to show."),
-                         default=('published', ),
-                         required=True,
-                         value_type=schema.Choice(
-                             vocabulary="plone.app.vocabularies.WorkflowStates")
-                         )
+        description=_(u"Items in which workflow state to show."),
+        default=('published', ),
+        required=True,
+        value_type=schema.Choice(
+            vocabulary="plone.app.vocabularies.WorkflowStates")
+        )
+
+    search_base = schema.Choice(
+        title=_(u'portlet_label_search_base', default=u'Search base'),
+        description=_(u'portlet_help_search_base',
+                      default=u'Select events search base folder'),
+        required=False,
+        source=SearchableTextSourceBinder({'is_folderish': True},
+                                           default_query='path:'),
+    )
+
 
 class Assignment(base.Assignment):
     implements(IEventsPortlet)
 
-    def __init__(self, count=5, state=('published', )):
+    # reduce upgrade pain
+    search_base = None
+
+    def __init__(self, count=5, state=('published', ), search_base=None):
         self.count = count
         self.state = state
+        self.search_base = search_base
 
     @property
     def title(self):
         return _(u"Events")
+
 
 class Renderer(base.Renderer):
 
@@ -119,13 +136,23 @@ class Renderer(base.Renderer):
     @memoize
     def _data(self):
         context = aq_inner(self.context)
-        limit = self.data.count
-        state = self.data.state
-        return get_portal_events(
+        data = self.data
+
+        query_kw = {}
+        if data.search_base:
+            query_kw['path'] = {'query': '%s%s' % (
+                '/'.join(self.portal.getPhysicalPath()), data.search_base)}
+
+        if data.state:
+            query_kw['review_state'] = data.state
+
+        ret = get_portal_events(
                 context,
                 range_start=localized_now(context),
-                limit=limit,
-                review_state=state)
+                limit=data.count,
+                **query_kw)
+
+        return ret
 
 class AddForm(base.AddForm):
     form_fields = form.Fields(IEventsPortlet)
@@ -133,7 +160,9 @@ class AddForm(base.AddForm):
     description = _(u"This portlet lists upcoming Events.")
 
     def create(self, data):
-        return Assignment(count=data.get('count', 5), state=data.get('state', ('published',)))
+        return Assignment(count=data.get('count', 5),
+                          state=data.get('state', ('published',)),
+                          search_base=data.get('search_base', 5))
 
 class EditForm(base.EditForm):
     form_fields = form.Fields(IEventsPortlet)